Reorganise into three separate files
